backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
passlib
bcrypt
python-jose
python-multipart
psycopg2-binary
pydantic[email]